Split federation.controllers to inbound and outbound
